Fix kau_gui 빌드: std::max(double, int) 추론 실패
width()/height() 는 int 라 1.0 과 그대로 섞으면 템플릿 인자 추론이 깨진다.
plot_strip.cpp 의 plot 사각형 계산 두 줄에 static_cast<double> 을 넣는다.

같은 패턴을 kau_gui 전체에서 확인했고 나머지 std::max/min/clamp 는
인자 타입이 일치한다.

+++ src/kau_gui/src/plot_strip.cpp
@@ -249,10 +249,11 @@ void PlotStrip::paintEvent(QPaintEvent *)
 
     p.fillRect(rect(), theme::PANEL_BG);
 
+    // width()/height() 는 int 다. 1.0 과 그대로 섞으면 std::max 추론이 깨진다.
     const QRectF plot(
         LEFT_PAD, TOP_PAD,
-        std::max(1.0, width() - LEFT_PAD - RIGHT_PAD),
-        std::max(1.0, height() - TOP_PAD - BOTTOM_PAD));
+        std::max(1.0, static_cast<double>(width() - LEFT_PAD - RIGHT_PAD)),
+        std::max(1.0, static_cast<double>(height() - TOP_PAD - BOTTOM_PAD)));
 
     double lo = 0.0;
